requirements.txt
## from valentine
# algorithms
numpy==1.26.0
pandas==2.1.1
nltk==3.8.1
anytree==2.10.0
networkx==3.1
chardet==5.2.0
jellyfish==1.0.1
PuLP==2.7.0
pyemd==1.0.0
# data loading
python-dateutil==2.8.2
# testing
pytest~=8.2.0
## added
torch==2.2.1
scipy~=1.11.3
scikit-learn~=1.3.1
transformers==4.37.2
datasets>=2.9.0
sentence-transformers
https://github.com/PythonOT/POT/archive/master.zip
tqdm~=4.66.1
plotly
umap-learn